+++ packages/gafl-webapp-service/src/uri.js
@@ -74,8 +74,10 @@ export const CANCEL_RP_AGREEMENT_NOT_FOUND = {
   uri: '/buy/cancel-recurring-payment/agreement-not-found',
   page: 'cancel-rp-agreement-not-found'
 }
-export const CANCEL_RP_LICENCE_NOT_FOUND = { uri: '/buy/cancel-recurring-payment/licence-not-found', page: 'licence-not-found' }
-export const CANCEL_RP_ALREADY_CANCELLED = { uri: '/buy/cancel-recurring-payment/already-cancelled', page: 'already-cancelled' }
+export const CANCEL_RP_LICENCE_NOT_FOUND = { uri: '/buy/cancel-recurring-payment/licence-not-found', page: 'cancel-rp-licence-not-found' }
+export const CANCEL_RP_ALREADY_CANCELLED = { uri: '/buy/cancel-recurring-payment/already-cancelled', page: 'cancel-rp-already-cancelled' }
+
+export const JOURNEY_GOAL = { uri: '/buy/journey-goal', page: 'journey-goal' }
 
 /**
  * These are informational static pages